requirements.txt
python-dotenv~=1.0.1
gspread~=6.1.4
pandas~=2.2.3
Unidecode~=1.3.8
numpy~=2.2.0